Git feedback
Make sure to always include a .gitginore file in your
projects. This is a plain text file that goes in the
root of your projects directory. It tells git which
files you would NOT like to check into the repo. Entries
are plain text name of either files or directories
you wish to ignore. Adding node_modules will keep
the node_modules/ folder out of the repo, which we
generally want to do because they are so large.