Node feedback
I was not able to get your app running due to a failed
require statement. It looks like node could not find
your key.js file, which means you probably forgot to
check it into the repository with git.

Also, when I cloned into the repo I got the node_modules
by default (rather than having to `npm install` them. Im
sure you know it was because of your `gitignore` files
naming. You were about 99% of the way there tho! Just
remember that the `gitignore` file is a configuration file
for git, and all configuration files begin with a .. So
instead of `_gitignore` you would name it `.gitignore` and
place it in the root of your directory. All entries are
just plain text names of either files or directory you
wish to ingore

Other than that, you code calling out to the apis looked
good.